Allow authentication via GitHub
Using OmniAuth we authenticate the user. If they are new we create them
a new account, however if they are a returning user we do not.

This is our first example of using a form object to extract the data
from the request and validate it. Then the operation takes the form
object and uses that to run the business logic. If the user is new then
we create a new account.

We use the Wisper gem to publish :success or :fail events that the
caller can subscribe to. In this case the caller is #create in our
SessionsController. This isn't a typical action but shows how we get the
data, validate it and run the logic.

Notice that we do not need to use strong parameters.