Introduce page objects to feature specs
After some research I wanted to try out Page Objects to encapsulate some
of the details about a page (or part of a page) and allow the feature
spec to be at a higher level. I think this is a nice way to remove
duplication and simplify the feature specs.

I'm still playing around with these but this is a good first step.