Refactor responsibilities of how a user session is managed
I wanted to refactor logging in, logging out and getting the current
user. Previously all of this was in the ApplicationController and it
felt a bit messy. Also I wanted the logging in to be part of the
LoginWithOmniAuth command.

First thing I did was to extract out the current user related code from
the ApplicationController and put it in the new Authentication::Warden
class. I struggled with a name for this but the dictionary definition is
"a person responsible for the supervision of a particular place or
activity or for enforcing the regulations associated with it", I also
considered Guard, Bouncer and some others but this felt ok for now. The
warden is passed a "store" so it can set and get the id of the current
user. As this is a web app, we just pass in the session, but this could
be any object that responds to [] and []=.

After that, I passed in the warden to the LoginWithOmniAuth which would
allow it to log the user in after authentication was successful.